+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 ---
 
-## 0.1.2
+## 0.1.7
 - Create application Ionic in blanck.
 - Created the "User" model.
+- Created the "Login" page.
+- Added the Facebook button on the "Login" page.
+- Created application developer on the Facebook.
+- Configured Ionic with Firebase and AngularFire.
+- Created AuthProvider to make the login with Firebase.